requirements.txt
streamlit
openai
python-dotenv
pandas
numpy
scikit-learn
faiss-cpu
matplotlib
plotly
tqdm
pydantic
